Consolidate LROs into one, and fixup PagedResponse name.
