+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keyvault_keys.hpp
@@ -13,10 +13,9 @@
 #include "azure/keyvault/keys/cryptography/cryptography_client_options.hpp"
 #include "azure/keyvault/keys/cryptography/decrypt_parameters.hpp"
 #include "azure/keyvault/keys/cryptography/encrypt_parameters.hpp"
-#include "azure/keyvault/keys/delete_key_operation.hpp"
 #include "azure/keyvault/keys/dll_import_export.hpp"
 #include "azure/keyvault/keys/key_client.hpp"
 #include "azure/keyvault/keys/key_client_models.hpp"
+#include "azure/keyvault/keys/key_client_operations.hpp"
 #include "azure/keyvault/keys/key_client_options.hpp"
-#include "azure/keyvault/keys/list_keys_responses.hpp"
-#include "azure/keyvault/keys/recover_deleted_key_operation.hpp"
+#include "azure/keyvault/keys/key_client_paged_responses.hpp"
