+++ sdk/keyvault/azure-security-keyvault-keys/src/delete_key_operation.cpp
@@ -5,7 +5,7 @@
 
 #include <azure/core/exception.hpp>
 
-#include "azure/keyvault/keys/delete_key_operation.hpp"
+#include "azure/keyvault/keys/key_client_operations.hpp"
 #include "azure/keyvault/keys/key_client.hpp"
 #include "private/key_constants.hpp"
 #include "private/key_serializers.hpp"
