+++ sdk/keyvault/azure-security-keyvault-keys/src/private/key_serializers.hpp
@@ -13,7 +13,7 @@
 
 #include "azure/keyvault/keys/key_client_models.hpp"
 #include "azure/keyvault/keys/key_client_options.hpp"
-#include "azure/keyvault/keys/list_keys_responses.hpp"
+#include "azure/keyvault/keys/key_client_paged_responses.hpp"
 
 #include <string>
