+++ sdk/keyvault/azure-security-keyvault-keys/src/recover_deleted_key_operation.cpp
@@ -4,7 +4,7 @@
 #include <azure/core/exception.hpp>
 
 #include "azure/keyvault/keys/key_client.hpp"
-#include "azure/keyvault/keys/recover_deleted_key_operation.hpp"
+#include "azure/keyvault/keys/key_client_operations.hpp"
 #include "private/key_constants.hpp"
 #include "private/key_serializers.hpp"
